# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Factory App API
  description: |
    REST API for the Industrial Operations Management System (Factory App).
    
    ## Authentication
    
    All API endpoints (except `/api/auth/login`) require a valid session cookie (`factory_session`).
    Mutating operations also require a valid CSRF token in the `X-CSRF-Token` header.
    
    ## Rate Limiting
    
    API endpoints are rate-limited:
    - General API: 100 requests per minute
    - Login: 5 attempts per minute
    - File uploads: 10 per minute
    
    When rate-limited, the API returns:
    - HTTP 429 (Too Many Requests)
    - `Retry-After` header with seconds until reset
    - `X-RateLimit-Remaining` header
    
    ## Error Responses
    
    All errors follow a consistent format:
    ```json
    {
      "error": "Human-readable error message",
      "code": "ERROR_CODE",
      "requestId": "req_abc123_xyz789",
      "timestamp": "2024-01-01T00:00:00.000Z"
    }
    ```
    
  version: 1.0.0
  contact:
    name: Factory App Development Team
  license:
    name: MIT
servers:
  - url: /api
    description: Local development
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with employee credentials
      description: Authenticates a user with their employee ID and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
        '403':
          description: Account locked
        '429':
          description: Too many login attempts
...
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      description: Returns the currently authenticated user's information
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Error code for programmatic handling
        requestId:
          type: string
          description: Unique request ID for debugging
        timestamp:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required:
        - employeeId
        - pin
      properties:
        employeeId:
          type: string
          minLength: 1
          example: "TECH-001"
        pin:
          type: string
          minLength: 4
          maxLength: 6
          example: "1234"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        csrfToken:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
        name:
          type: string
        roleName:
          type: string
        permissions:
          type: array
          items:
            type: string
    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        csrfToken:
          type: string
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: factory_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
security:
  - sessionCookie: []